requirements.txt
pydantic-ai-slim[openai]
pydantic-ai-slim[groq]
python_dotenv
torchaudio
omegaconf